@@ -54,12 +54,17 @@ func (b *BaseChain) EthereumTypeEstimateGas(params map[string]interface{}) (uint
 	return 0, errors.New("Not supported")
 }
 
-// EthereumTypeGetErc20ContractInfo is not supported
-func (b *BaseChain) EthereumTypeGetErc20ContractInfo(contractDesc AddressDescriptor) (*Erc20Contract, error) {
+// GetContractInfo is not supported
+func (b *BaseChain) GetContractInfo(contractDesc AddressDescriptor) (*ContractInfo, error) {
 	return nil, errors.New("Not supported")
 }
 
 // EthereumTypeGetErc20ContractBalance is not supported
 func (b *BaseChain) EthereumTypeGetErc20ContractBalance(addrDesc, contractDesc AddressDescriptor) (*big.Int, error) {
 	return nil, errors.New("Not supported")
+}
+
+// GetContractInfo returns URI of non fungible or multi token defined by token id
+func (p *BaseChain) GetTokenURI(contractDesc AddressDescriptor, tokenID *big.Int) (string, error) {
+	return "", errors.New("Not supported")
 }
