@@ -16,9 +16,11 @@ type BitcoinParser struct {
 
 // NewBitcoinParser returns new BitcoinParser instance
 func NewBitcoinParser(params *chaincfg.Params, c *Configuration) *BitcoinParser {
-	return &BitcoinParser{
+	p := &BitcoinParser{
 		BitcoinLikeParser: NewBitcoinLikeParser(params, c),
 	}
+	p.VSizeSupport = true
+	return p
 }
 
 // GetChainParams contains network parameters for the main Bitcoin network,
@@ -63,6 +65,7 @@ type Tx struct {
 	Txid        string       `json:"txid"`
 	Version     int32        `json:"version"`
 	LockTime    uint32       `json:"locktime"`
+	VSize       int64        `json:"vsize,omitempty"`
 	Vin         []bchain.Vin `json:"vin"`
 	Vout        []Vout       `json:"vout"`
 	BlockHeight uint32       `json:"blockHeight,omitempty"`
@@ -88,6 +91,7 @@ func (p *BitcoinParser) ParseTxFromJson(msg json.RawMessage) (*bchain.Tx, error)
 	tx.Txid = bitcoinTx.Txid
 	tx.Version = bitcoinTx.Version
 	tx.LockTime = bitcoinTx.LockTime
+	tx.VSize = bitcoinTx.VSize
 	tx.Vin = bitcoinTx.Vin
 	tx.BlockHeight = bitcoinTx.BlockHeight
 	tx.Confirmations = bitcoinTx.Confirmations
