@@ -34,7 +34,9 @@ type NamecoinParser struct {
 
 // NewNamecoinParser returns new NamecoinParser instance
 func NewNamecoinParser(params *chaincfg.Params, c *btc.Configuration) *NamecoinParser {
-	return &NamecoinParser{BitcoinLikeParser: btc.NewBitcoinLikeParser(params, c)}
+	p := &NamecoinParser{BitcoinLikeParser: btc.NewBitcoinLikeParser(params, c)}
+	p.VSizeSupport = true
+	return p
 }
 
 // GetChainParams contains network parameters for the main Namecoin network,
