@@ -4,7 +4,10 @@ import (
 	"encoding/json"
 	"sort"
 	"sync"
+	"sync/atomic"
 	"time"
+
+	"github.com/golang/glog"
 )
 
 const (
@@ -15,6 +18,8 @@ const (
 	// DbStateInconsistent means db is in inconsistent state and cannot be used
 	DbStateInconsistent
 )
+
+var inShutdown int32
 
 // InternalStateColumn contains the data of a db column
 type InternalStateColumn struct {
@@ -28,19 +33,20 @@ type InternalStateColumn struct {
 
 // BackendInfo is used to get information about blockchain
 type BackendInfo struct {
-	BackendError    string      `json:"error,omitempty"`
-	Chain           string      `json:"chain,omitempty"`
-	Blocks          int         `json:"blocks,omitempty"`
-	Headers         int         `json:"headers,omitempty"`
-	BestBlockHash   string      `json:"bestBlockHash,omitempty"`
-	Difficulty      string      `json:"difficulty,omitempty"`
-	SizeOnDisk      int64       `json:"sizeOnDisk,omitempty"`
-	Version         string      `json:"version,omitempty"`
-	Subversion      string      `json:"subversion,omitempty"`
-	ProtocolVersion string      `json:"protocolVersion,omitempty"`
-	Timeoffset      float64     `json:"timeOffset,omitempty"`
-	Warnings        string      `json:"warnings,omitempty"`
-	Consensus       interface{} `json:"consensus,omitempty"`
+	BackendError     string      `json:"error,omitempty"`
+	Chain            string      `json:"chain,omitempty"`
+	Blocks           int         `json:"blocks,omitempty"`
+	Headers          int         `json:"headers,omitempty"`
+	BestBlockHash    string      `json:"bestBlockHash,omitempty"`
+	Difficulty       string      `json:"difficulty,omitempty"`
+	SizeOnDisk       int64       `json:"sizeOnDisk,omitempty"`
+	Version          string      `json:"version,omitempty"`
+	Subversion       string      `json:"subversion,omitempty"`
+	ProtocolVersion  string      `json:"protocolVersion,omitempty"`
+	Timeoffset       float64     `json:"timeOffset,omitempty"`
+	Warnings         string      `json:"warnings,omitempty"`
+	ConsensusVersion string      `json:"consensus_version,omitempty"`
+	Consensus        interface{} `json:"consensus,omitempty"`
 }
 
 // InternalState contains the data of the internal state
@@ -52,7 +58,8 @@ type InternalState struct {
 	CoinLabel    string `json:"coinLabel"`
 	Host         string `json:"host"`
 
-	DbState uint32 `json:"dbState"`
+	DbState       uint32 `json:"dbState"`
+	ExtendedIndex bool   `json:"extendedIndex"`
 
 	LastStore time.Time `json:"lastStore"`
 
@@ -62,8 +69,10 @@ type InternalState struct {
 	InitialSync    bool      `json:"initialSync"`
 	IsSynchronized bool      `json:"isSynchronized"`
 	BestHeight     uint32    `json:"bestHeight"`
+	StartSync      time.Time `json:"-"`
 	LastSync       time.Time `json:"lastSync"`
 	BlockTimes     []uint32  `json:"-"`
+	AvgBlockPeriod uint32    `json:"-"`
 
 	IsMempoolSynchronized bool      `json:"isMempoolSynchronized"`
 	MempoolSize           int       `json:"mempoolSize"`
@@ -71,15 +80,25 @@ type InternalState struct {
 
 	DbColumns []InternalStateColumn `json:"dbColumns"`
 
-	UtxoChecked bool `json:"utxoChecked"`
+	HasFiatRates                 bool      `json:"-"`
+	HasTokenFiatRates            bool      `json:"-"`
+	HistoricalFiatRatesTime      time.Time `json:"historicalFiatRatesTime"`
+	HistoricalTokenFiatRatesTime time.Time `json:"historicalTokenFiatRatesTime"`
+
+	EnableSubNewTx bool `json:"-"`
 
 	BackendInfo BackendInfo `json:"-"`
+
+	// database migrations
+	UtxoChecked            bool `json:"utxoChecked"`
+	SortedAddressContracts bool `json:"sortedAddressContracts"`
 }
 
 // StartedSync signals start of synchronization
 func (is *InternalState) StartedSync() {
 	is.mux.Lock()
 	defer is.mux.Unlock()
+	is.StartSync = time.Now().UTC()
 	is.IsSynchronized = false
 }
 
@@ -89,7 +108,7 @@ func (is *InternalState) FinishedSync(bestHeight uint32) {
 	defer is.mux.Unlock()
 	is.IsSynchronized = true
 	is.BestHeight = bestHeight
-	is.LastSync = time.Now()
+	is.LastSync = time.Now().UTC()
 }
 
 // UpdateBestHeight sets new best height, without changing IsSynchronized flag
@@ -97,7 +116,7 @@ func (is *InternalState) UpdateBestHeight(bestHeight uint32) {
 	is.mux.Lock()
 	defer is.mux.Unlock()
 	is.BestHeight = bestHeight
-	is.LastSync = time.Now()
+	is.LastSync = time.Now().UTC()
 }
 
 // FinishedSyncNoChange marks end of synchronization in case no index update was necessary, it does not update lastSync time
@@ -108,10 +127,10 @@ func (is *InternalState) FinishedSyncNoChange() {
 }
 
 // GetSyncState gets the state of synchronization
-func (is *InternalState) GetSyncState() (bool, uint32, time.Time) {
+func (is *InternalState) GetSyncState() (bool, uint32, time.Time, time.Time) {
 	is.mux.Lock()
 	defer is.mux.Unlock()
-	return is.IsSynchronized, is.BestHeight, is.LastSync
+	return is.IsSynchronized, is.BestHeight, is.LastSync, is.StartSync
 }
 
 // StartedMempoolSync signals start of mempool synchronization
@@ -197,11 +216,33 @@ func (is *InternalState) GetBlockTime(height uint32) uint32 {
 	return 0
 }
 
-// AppendBlockTime appends block time to BlockTimes
-func (is *InternalState) AppendBlockTime(time uint32) {
+// GetLastBlockTime returns time of the last block
+func (is *InternalState) GetLastBlockTime() uint32 {
+	is.mux.Lock()
+	defer is.mux.Unlock()
+	if len(is.BlockTimes) > 0 {
+		return is.BlockTimes[len(is.BlockTimes)-1]
+	}
+	return 0
+}
+
+// SetBlockTimes initializes BlockTimes array, returns AvgBlockPeriod
+func (is *InternalState) SetBlockTimes(blockTimes []uint32) uint32 {
+	is.mux.Lock()
+	defer is.mux.Unlock()
+	is.BlockTimes = blockTimes
+	is.computeAvgBlockPeriod()
+	glog.Info("set ", len(is.BlockTimes), " block times, average block period ", is.AvgBlockPeriod, "s")
+	return is.AvgBlockPeriod
+}
+
+// AppendBlockTime appends block time to BlockTimes, returns AvgBlockPeriod
+func (is *InternalState) AppendBlockTime(time uint32) uint32 {
 	is.mux.Lock()
 	defer is.mux.Unlock()
 	is.BlockTimes = append(is.BlockTimes, time)
+	is.computeAvgBlockPeriod()
+	return is.AvgBlockPeriod
 }
 
 // RemoveLastBlockTimes removes last times from BlockTimes
@@ -212,6 +253,7 @@ func (is *InternalState) RemoveLastBlockTimes(count int) {
 		count = len(is.BlockTimes)
 	}
 	is.BlockTimes = is.BlockTimes[:len(is.BlockTimes)-count]
+	is.computeAvgBlockPeriod()
 }
 
 // GetBlockHeightOfTime returns block height of the first block with time greater or equal to the given time or MaxUint32 if no such block
@@ -233,6 +275,25 @@ func (is *InternalState) GetBlockHeightOfTime(time uint32) uint32 {
 		}
 	}
 	return uint32(height)
+}
+
+const avgBlockPeriodSample = 100
+
+// Avg100BlocksPeriod returns average period of the last 100 blocks in seconds
+func (is *InternalState) GetAvgBlockPeriod() uint32 {
+	is.mux.Lock()
+	defer is.mux.Unlock()
+	return is.AvgBlockPeriod
+}
+
+// computeAvgBlockPeriod returns computes average of the last 100 blocks in seconds
+func (is *InternalState) computeAvgBlockPeriod() {
+	last := len(is.BlockTimes) - 1
+	first := last - avgBlockPeriodSample - 1
+	if first < 0 {
+		return
+	}
+	is.AvgBlockPeriod = (is.BlockTimes[last] - is.BlockTimes[first]) / avgBlockPeriodSample
 }
 
 // SetBackendInfo sets new BackendInfo
@@ -264,4 +325,14 @@ func UnpackInternalState(buf []byte) (*InternalState, error) {
 		return nil, err
 	}
 	return &is, nil
+}
+
+// SetInShutdown sets the internal state to in shutdown state
+func SetInShutdown() {
+	atomic.StoreInt32(&inShutdown, 1)
+}
+
+// IsInShutdown returns true if in application shutdown state
+func IsInShutdown() bool {
+	return atomic.LoadInt32(&inShutdown) != 0
 }
