@@ -3,8 +3,8 @@ package db
 import (
 	"time"
 
-	"github.com/flier/gorocksdb"
 	"github.com/golang/glog"
+	"github.com/linxGnu/grocksdb"
 	"github.com/trezor/blockbook/bchain"
 )
 
@@ -25,6 +25,7 @@ type BulkConnect struct {
 	chainType          bchain.ChainType
 	bulkAddresses      []bulkAddresses
 	bulkAddressesCount int
+	ethBlockTxs        []ethBlockTx
 	txAddressesMap     map[string]*TxAddresses
 	balances           map[string]*AddrBalance
 	addressContracts   map[string]*AddrContracts
@@ -57,7 +58,7 @@ func (d *RocksDB) InitBulkConnect() (*BulkConnect, error) {
 	return b, nil
 }
 
-func (b *BulkConnect) storeTxAddresses(wb *gorocksdb.WriteBatch, all bool) (int, int, error) {
+func (b *BulkConnect) storeTxAddresses(wb *grocksdb.WriteBatch, all bool) (int, int, error) {
 	var txm map[string]*TxAddresses
 	var sp int
 	if all {
@@ -100,14 +101,14 @@ func (b *BulkConnect) storeTxAddresses(wb *gorocksdb.WriteBatch, all bool) (int,
 func (b *BulkConnect) parallelStoreTxAddresses(c chan error, all bool) {
 	defer close(c)
 	start := time.Now()
-	wb := gorocksdb.NewWriteBatch()
+	wb := grocksdb.NewWriteBatch()
 	defer wb.Destroy()
 	count, sp, err := b.storeTxAddresses(wb, all)
 	if err != nil {
 		c <- err
 		return
 	}
-	if err := b.d.db.Write(b.d.wo, wb); err != nil {
+	if err := b.d.WriteBatch(wb); err != nil {
 		c <- err
 		return
 	}
@@ -115,7 +116,7 @@ func (b *BulkConnect) parallelStoreTxAddresses(c chan error, all bool) {
 	c <- nil
 }
 
-func (b *BulkConnect) storeBalances(wb *gorocksdb.WriteBatch, all bool) (int, error) {
+func (b *BulkConnect) storeBalances(wb *grocksdb.WriteBatch, all bool) (int, error) {
 	var bal map[string]*AddrBalance
 	if all {
 		bal = b.balances
@@ -140,14 +141,14 @@ func (b *BulkConnect) storeBalances(wb *gorocksdb.WriteBatch, all bool) (int, er
 func (b *BulkConnect) parallelStoreBalances(c chan error, all bool) {
 	defer close(c)
 	start := time.Now()
-	wb := gorocksdb.NewWriteBatch()
+	wb := grocksdb.NewWriteBatch()
 	defer wb.Destroy()
 	count, err := b.storeBalances(wb, all)
 	if err != nil {
 		c <- err
 		return
 	}
-	if err := b.d.db.Write(b.d.wo, wb); err != nil {
+	if err := b.d.WriteBatch(wb); err != nil {
 		c <- err
 		return
 	}
@@ -155,7 +156,7 @@ func (b *BulkConnect) parallelStoreBalances(c chan error, all bool) {
 	c <- nil
 }
 
-func (b *BulkConnect) storeBulkAddresses(wb *gorocksdb.WriteBatch) error {
+func (b *BulkConnect) storeBulkAddresses(wb *grocksdb.WriteBatch) error {
 	for _, ba := range b.bulkAddresses {
 		if err := b.d.storeAddresses(wb, ba.bi.Height, ba.addresses); err != nil {
 			return err
@@ -201,7 +202,7 @@ func (b *BulkConnect) connectBlockBitcoinType(block *bchain.Block, storeBlockTxs
 	// open WriteBatch only if going to write
 	if sa || b.bulkAddressesCount > maxBulkAddresses || storeBlockTxs {
 		start := time.Now()
-		wb := gorocksdb.NewWriteBatch()
+		wb := grocksdb.NewWriteBatch()
 		defer wb.Destroy()
 		bac := b.bulkAddressesCount
 		if sa || b.bulkAddressesCount > maxBulkAddresses {
@@ -214,7 +215,7 @@ func (b *BulkConnect) connectBlockBitcoinType(block *bchain.Block, storeBlockTxs
 				return err
 			}
 		}
-		if err := b.d.db.Write(b.d.wo, wb); err != nil {
+		if err := b.d.WriteBatch(wb); err != nil {
 			return err
 		}
 		if bac > b.bulkAddressesCount {
@@ -234,7 +235,7 @@ func (b *BulkConnect) connectBlockBitcoinType(block *bchain.Block, storeBlockTxs
 	return nil
 }
 
-func (b *BulkConnect) storeAddressContracts(wb *gorocksdb.WriteBatch, all bool) (int, error) {
+func (b *BulkConnect) storeAddressContracts(wb *grocksdb.WriteBatch, all bool) (int, error) {
 	var ac map[string]*AddrContracts
 	if all {
 		ac = b.addressContracts
@@ -259,14 +260,14 @@ func (b *BulkConnect) storeAddressContracts(wb *gorocksdb.WriteBatch, all bool) 
 func (b *BulkConnect) parallelStoreAddressContracts(c chan error, all bool) {
 	defer close(c)
 	start := time.Now()
-	wb := gorocksdb.NewWriteBatch()
+	wb := grocksdb.NewWriteBatch()
 	defer wb.Destroy()
 	count, err := b.storeAddressContracts(wb, all)
 	if err != nil {
 		c <- err
 		return
 	}
-	if err := b.d.db.Write(b.d.wo, wb); err != nil {
+	if err := b.d.WriteBatch(wb); err != nil {
 		c <- err
 		return
 	}
@@ -280,6 +281,7 @@ func (b *BulkConnect) connectBlockEthereumType(block *bchain.Block, storeBlockTx
 	if err != nil {
 		return err
 	}
+	b.ethBlockTxs = append(b.ethBlockTxs, blockTxs...)
 	var storeAddrContracts chan error
 	var sa bool
 	if len(b.addressContracts) > maxBulkAddrContracts {
@@ -301,24 +303,44 @@ func (b *BulkConnect) connectBlockEthereumType(block *bchain.Block, storeBlockTx
 	// open WriteBatch only if going to write
 	if sa || b.bulkAddressesCount > maxBulkAddresses || storeBlockTxs {
 		start := time.Now()
-		wb := gorocksdb.NewWriteBatch()
+		wb := grocksdb.NewWriteBatch()
 		defer wb.Destroy()
 		bac := b.bulkAddressesCount
 		if sa || b.bulkAddressesCount > maxBulkAddresses {
-			if err := b.storeBulkAddresses(wb); err != nil {
+			if err = b.storeBulkAddresses(wb); err != nil {
 				return err
 			}
+		}
+		if err = b.d.storeInternalDataEthereumType(wb, b.ethBlockTxs); err != nil {
+			return err
+		}
+		b.ethBlockTxs = b.ethBlockTxs[:0]
+		if err = b.d.storeBlockSpecificDataEthereumType(wb, block); err != nil {
+			return err
 		}
 		if storeBlockTxs {
-			if err := b.d.storeAndCleanupBlockTxsEthereumType(wb, block, blockTxs); err != nil {
+			if err = b.d.storeAndCleanupBlockTxsEthereumType(wb, block, blockTxs); err != nil {
 				return err
 			}
 		}
-		if err := b.d.db.Write(b.d.wo, wb); err != nil {
+		if err = b.d.WriteBatch(wb); err != nil {
 			return err
 		}
 		if bac > b.bulkAddressesCount {
 			glog.Info("rocksdb: height ", b.height, ", stored ", bac, " addresses, done in ", time.Since(start))
+		}
+	} else {
+		// if there are blockSpecificData, store them
+		blockSpecificData, _ := block.CoinSpecificData.(*bchain.EthereumBlockSpecificData)
+		if blockSpecificData != nil {
+			wb := grocksdb.NewWriteBatch()
+			defer wb.Destroy()
+			if err = b.d.storeBlockSpecificDataEthereumType(wb, block); err != nil {
+				return err
+			}
+			if err := b.d.WriteBatch(wb); err != nil {
+				return err
+			}
 		}
 	}
 	if storeAddrContracts != nil {
@@ -356,13 +378,17 @@ func (b *BulkConnect) Close() error {
 		storeAddressContractsChan = make(chan error)
 		go b.parallelStoreAddressContracts(storeAddressContractsChan, true)
 	}
-	wb := gorocksdb.NewWriteBatch()
+	wb := grocksdb.NewWriteBatch()
 	defer wb.Destroy()
+	if err := b.d.storeInternalDataEthereumType(wb, b.ethBlockTxs); err != nil {
+		return err
+	}
+	b.ethBlockTxs = b.ethBlockTxs[:0]
 	bac := b.bulkAddressesCount
 	if err := b.storeBulkAddresses(wb); err != nil {
 		return err
 	}
-	if err := b.d.db.Write(b.d.wo, wb); err != nil {
+	if err := b.d.WriteBatch(wb); err != nil {
 		return err
 	}
 	glog.Info("rocksdb: height ", b.height, ", stored ", bac, " addresses, done in ", time.Since(start))
@@ -381,10 +407,13 @@ func (b *BulkConnect) Close() error {
 			return err
 		}
 	}
-	var err error
-	b.d.is.BlockTimes, err = b.d.loadBlockTimes()
+	bt, err := b.d.loadBlockTimes()
 	if err != nil {
 		return err
+	}
+	avg := b.d.is.SetBlockTimes(bt)
+	if b.d.metrics != nil {
+		b.d.metrics.AvgBlockPeriod.Set(float64(avg))
 	}
 
 	if err := b.d.SetInconsistentState(false); err != nil {
