@@ -2,31 +2,28 @@ package db
 
 // #include "rocksdb/c.h"
 import "C"
-
-import (
-	"github.com/flier/gorocksdb"
-)
+import "github.com/linxGnu/grocksdb"
 
 /*
-	possible additional tuning, using options not accessible by gorocksdb
+	possible additional tuning, using options not accessible by grocksdb
 
 // #include "rocksdb/c.h"
 import "C"
 
 	cNativeOpts := C.rocksdb_options_create()
-	opts := &gorocksdb.Options{}
+	opts := &grocksdb.Options{}
 	cField := reflect.Indirect(reflect.ValueOf(opts)).FieldByName("c")
 	cPtr := (**C.rocksdb_options_t)(unsafe.Pointer(cField.UnsafeAddr()))
 	*cPtr = cNativeOpts
 
 	cNativeBlockOpts := C.rocksdb_block_based_options_create()
-	blockOpts := &gorocksdb.BlockBasedTableOptions{}
+	blockOpts := &grocksdb.BlockBasedTableOptions{}
 	cBlockField := reflect.Indirect(reflect.ValueOf(blockOpts)).FieldByName("c")
 	cBlockPtr := (**C.rocksdb_block_based_table_options_t)(unsafe.Pointer(cBlockField.UnsafeAddr()))
 	*cBlockPtr = cNativeBlockOpts
 
 	// https://github.com/facebook/rocksdb/wiki/Partitioned-Index-Filters
-	blockOpts.SetIndexType(gorocksdb.KTwoLevelIndexSearchIndexType)
+	blockOpts.SetIndexType(grocksdb.KTwoLevelIndexSearchIndexType)
 	C.rocksdb_block_based_options_set_partition_filters(cNativeBlockOpts, boolToChar(true))
 	C.rocksdb_block_based_options_set_metadata_block_size(cNativeBlockOpts, C.uint64_t(4096))
 	C.rocksdb_block_based_options_set_cache_index_and_filter_blocks_with_high_priority(cNativeBlockOpts, boolToChar(true))
@@ -41,16 +38,16 @@ func boolToChar(b bool) C.uchar {
 }
 */
 
-func createAndSetDBOptions(bloomBits int, c *gorocksdb.Cache, maxOpenFiles int) *gorocksdb.Options {
-	blockOpts := gorocksdb.NewDefaultBlockBasedTableOptions()
+func createAndSetDBOptions(bloomBits int, c *grocksdb.Cache, maxOpenFiles int) *grocksdb.Options {
+	blockOpts := grocksdb.NewDefaultBlockBasedTableOptions()
 	blockOpts.SetBlockSize(32 << 10) // 32kB
 	blockOpts.SetBlockCache(c)
 	if bloomBits > 0 {
-		blockOpts.SetFilterPolicy(gorocksdb.NewBloomFilter(bloomBits))
+		blockOpts.SetFilterPolicy(grocksdb.NewBloomFilter(float64(bloomBits)))
 	}
 	blockOpts.SetFormatVersion(4)
 
-	opts := gorocksdb.NewDefaultOptions()
+	opts := grocksdb.NewDefaultOptions()
 	opts.SetBlockBasedTableFactory(blockOpts)
 	opts.SetCreateIfMissing(true)
 	opts.SetCreateIfMissingColumnFamilies(true)
@@ -60,6 +57,6 @@ func createAndSetDBOptions(bloomBits int, c *gorocksdb.Cache, maxOpenFiles int) 
 	opts.SetWriteBufferSize(1 << 27)      // 128MB
 	opts.SetMaxBytesForLevelBase(1 << 27) // 128MB
 	opts.SetMaxOpenFiles(maxOpenFiles)
-	opts.SetCompression(gorocksdb.LZ4HCCompression)
+	opts.SetCompression(grocksdb.LZ4HCCompression)
 	return opts
 }
