@@ -18,6 +18,7 @@ import (
 	"github.com/trezor/blockbook/bchain"
 	"github.com/trezor/blockbook/common"
 	"github.com/trezor/blockbook/db"
+	"github.com/trezor/blockbook/fiat"
 )
 
 const upgradeFailed = "Upgrade failed: "
@@ -34,21 +35,10 @@ var (
 	connectionCounter uint64
 )
 
-type websocketReq struct {
-	ID     string          `json:"id"`
-	Method string          `json:"method"`
-	Params json.RawMessage `json:"params"`
-}
-
-type websocketRes struct {
-	ID   string      `json:"id"`
-	Data interface{} `json:"data"`
-}
-
 type websocketChannel struct {
 	id            uint64
 	conn          *websocket.Conn
-	out           chan *websocketRes
+	out           chan *WsRes
 	ip            string
 	requestHeader http.Header
 	alive         bool
@@ -58,7 +48,6 @@ type websocketChannel struct {
 
 // WebsocketServer is a handle to websocket server
 type WebsocketServer struct {
-	socket                          *websocket.Conn
 	upgrader                        *websocket.Upgrader
 	db                              *db.RocksDB
 	txCache                         *db.TxCache
@@ -77,12 +66,13 @@ type WebsocketServer struct {
 	addressSubscriptions            map[string]map[*websocketChannel]string
 	addressSubscriptionsLock        sync.Mutex
 	fiatRatesSubscriptions          map[string]map[*websocketChannel]string
+	fiatRatesTokenSubscriptions     map[*websocketChannel][]string
 	fiatRatesSubscriptionsLock      sync.Mutex
 }
 
 // NewWebsocketServer creates new websocket interface to blockbook and returns its handle
-func NewWebsocketServer(db *db.RocksDB, chain bchain.BlockChain, mempool bchain.Mempool, txCache *db.TxCache, metrics *common.Metrics, is *common.InternalState, enableSubNewTx bool) (*WebsocketServer, error) {
-	api, err := api.NewWorker(db, chain, mempool, txCache, metrics, is)
+func NewWebsocketServer(db *db.RocksDB, chain bchain.BlockChain, mempool bchain.Mempool, txCache *db.TxCache, metrics *common.Metrics, is *common.InternalState, fiatRates *fiat.FiatRates) (*WebsocketServer, error) {
+	api, err := api.NewWorker(db, chain, mempool, txCache, metrics, is, fiatRates)
 	if err != nil {
 		return nil, err
 	}
@@ -106,10 +96,11 @@ func NewWebsocketServer(db *db.RocksDB, chain bchain.BlockChain, mempool bchain.
 		api:                         api,
 		block0hash:                  b0,
 		newBlockSubscriptions:       make(map[*websocketChannel]string),
-		newTransactionEnabled:       enableSubNewTx,
+		newTransactionEnabled:       is.EnableSubNewTx,
 		newTransactionSubscriptions: make(map[*websocketChannel]string),
 		addressSubscriptions:        make(map[string]map[*websocketChannel]string),
 		fiatRatesSubscriptions:      make(map[string]map[*websocketChannel]string),
+		fiatRatesTokenSubscriptions: make(map[*websocketChannel][]string),
 	}
 	return s, nil
 }
@@ -141,7 +132,7 @@ func (s *WebsocketServer) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	c := &websocketChannel{
 		id:            atomic.AddUint64(&connectionCounter, 1),
 		conn:          conn,
-		out:           make(chan *websocketRes, outChannelSize),
+		out:           make(chan *WsRes, outChannelSize),
 		ip:            getIP(r),
 		requestHeader: r.Header,
 		alive:         true,
@@ -178,7 +169,7 @@ func (c *websocketChannel) CloseOut() bool {
 	return false
 }
 
-func (c *websocketChannel) DataOut(data *websocketRes) {
+func (c *websocketChannel) DataOut(data *WsRes) {
 	c.aliveLock.Lock()
 	defer c.aliveLock.Unlock()
 	if c.alive {
@@ -209,7 +200,7 @@ func (s *WebsocketServer) inputLoop(c *websocketChannel) {
 		}
 		switch t {
 		case websocket.TextMessage:
-			var req websocketReq
+			var req WsReq
 			err := json.Unmarshal(d, &req)
 			if err != nil {
 				glog.Error("Error parsing message from ", c.id, ", ", string(d), ", ", err)
@@ -223,7 +214,6 @@ func (s *WebsocketServer) inputLoop(c *websocketChannel) {
 			return
 		case websocket.PingMessage:
 			c.conn.WriteControl(websocket.PongMessage, nil, time.Now().Add(defaultTimeout))
-			break
 		case websocket.CloseMessage:
 			s.closeChannel(c)
 			return
@@ -264,46 +254,49 @@ func (s *WebsocketServer) onDisconnect(c *websocketChannel) {
 	s.metrics.WebsocketClients.Dec()
 }
 
-var requestHandlers = map[string]func(*WebsocketServer, *websocketChannel, *websocketReq) (interface{}, error){
-	"getAccountInfo": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
+var requestHandlers = map[string]func(*WebsocketServer, *websocketChannel, *WsReq) (interface{}, error){
+	"getAccountInfo": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
 		r, err := unmarshalGetAccountInfoRequest(req.Params)
 		if err == nil {
 			rv, err = s.getAccountInfo(r)
 		}
 		return
 	},
-	"getInfo": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
+	"getInfo": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
 		return s.getInfo()
 	},
-	"getBlockHash": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
-		r := struct {
-			Height int `json:"height"`
-		}{}
+	"getBlockHash": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
+		r := WsBlockHashReq{}
 		err = json.Unmarshal(req.Params, &r)
 		if err == nil {
 			rv, err = s.getBlockHash(r.Height)
 		}
 		return
 	},
-	"getAccountUtxo": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
-		r := struct {
-			Descriptor string `json:"descriptor"`
-		}{}
+	"getBlock": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
+		if !s.is.ExtendedIndex {
+			return nil, errors.New("Not supported")
+		}
+		r := WsBlockReq{}
+		err = json.Unmarshal(req.Params, &r)
+		if r.PageSize == 0 {
+			r.PageSize = 1000000
+		}
+		if err == nil {
+			rv, err = s.getBlock(r.Id, r.Page, r.PageSize)
+		}
+		return
+	},
+	"getAccountUtxo": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
+		r := WsAccountUtxoReq{}
 		err = json.Unmarshal(req.Params, &r)
 		if err == nil {
 			rv, err = s.getAccountUtxo(r.Descriptor)
 		}
 		return
 	},
-	"getBalanceHistory": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
-		r := struct {
-			Descriptor string   `json:"descriptor"`
-			From       int64    `json:"from"`
-			To         int64    `json:"to"`
-			Currencies []string `json:"currencies"`
-			Gap        int      `json:"gap"`
-			GroupBy    uint32   `json:"groupBy"`
-		}{}
+	"getBalanceHistory": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
+		r := WsBalanceHistoryReq{}
 		err = json.Unmarshal(req.Params, &r)
 		if err == nil {
 			if r.From <= 0 {
@@ -322,112 +315,109 @@ var requestHandlers = map[string]func(*WebsocketServer, *websocketChannel, *webs
 		}
 		return
 	},
-	"getTransaction": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
-		r := struct {
-			Txid string `json:"txid"`
-		}{}
+	"getTransaction": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
+		r := WsTransactionReq{}
 		err = json.Unmarshal(req.Params, &r)
 		if err == nil {
 			rv, err = s.getTransaction(r.Txid)
 		}
 		return
 	},
-	"getTransactionSpecific": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
-		r := struct {
-			Txid string `json:"txid"`
-		}{}
+	"getTransactionSpecific": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
+		r := WsTransactionSpecificReq{}
 		err = json.Unmarshal(req.Params, &r)
 		if err == nil {
 			rv, err = s.getTransactionSpecific(r.Txid)
 		}
 		return
 	},
-	"estimateFee": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
+	"estimateFee": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
 		return s.estimateFee(c, req.Params)
 	},
-	"sendTransaction": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
-		r := struct {
-			Hex string `json:"hex"`
-		}{}
+	"sendTransaction": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
+		r := WsSendTransactionReq{}
 		err = json.Unmarshal(req.Params, &r)
 		if err == nil {
 			rv, err = s.sendTransaction(r.Hex)
 		}
 		return
 	},
-	"subscribeNewBlock": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
+	"getMempoolFilters": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
+		r := WsMempoolFiltersReq{}
+		err = json.Unmarshal(req.Params, &r)
+		if err == nil {
+			rv, err = s.getMempoolFilters(&r)
+		}
+		return
+	},
+	"subscribeNewBlock": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
 		return s.subscribeNewBlock(c, req)
 	},
-	"unsubscribeNewBlock": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
+	"unsubscribeNewBlock": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
 		return s.unsubscribeNewBlock(c)
 	},
-	"subscribeNewTransaction": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
+	"subscribeNewTransaction": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
 		return s.subscribeNewTransaction(c, req)
 	},
-	"unsubscribeNewTransaction": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
+	"unsubscribeNewTransaction": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
 		return s.unsubscribeNewTransaction(c)
 	},
-	"subscribeAddresses": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
+	"subscribeAddresses": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
 		ad, err := s.unmarshalAddresses(req.Params)
 		if err == nil {
 			rv, err = s.subscribeAddresses(c, ad, req)
 		}
 		return
 	},
-	"unsubscribeAddresses": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
+	"unsubscribeAddresses": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
 		return s.unsubscribeAddresses(c)
 	},
-	"subscribeFiatRates": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
-		r := struct {
-			Currency string `json:"currency"`
-		}{}
+	"subscribeFiatRates": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
+		var r WsSubscribeFiatRatesReq
 		err = json.Unmarshal(req.Params, &r)
 		if err != nil {
 			return nil, err
 		}
-		return s.subscribeFiatRates(c, strings.ToLower(r.Currency), req)
+		r.Currency = strings.ToLower(r.Currency)
+		for i := range r.Tokens {
+			r.Tokens[i] = strings.ToLower(r.Tokens[i])
+		}
+		return s.subscribeFiatRates(c, &r, req)
 	},
-	"unsubscribeFiatRates": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
+	"unsubscribeFiatRates": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
 		return s.unsubscribeFiatRates(c)
 	},
-	"ping": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
+	"ping": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
 		r := struct{}{}
 		return r, nil
 	},
-	"getCurrentFiatRates": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
-		r := struct {
-			Currencies []string `json:"currencies"`
-		}{}
+	"getCurrentFiatRates": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
+		r := WsCurrentFiatRatesReq{}
 		err = json.Unmarshal(req.Params, &r)
 		if err == nil {
-			rv, err = s.getCurrentFiatRates(r.Currencies)
+			rv, err = s.getCurrentFiatRates(r.Currencies, r.Token)
 		}
 		return
 	},
-	"getFiatRatesForTimestamps": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
-		r := struct {
-			Timestamps []int64  `json:"timestamps"`
-			Currencies []string `json:"currencies"`
-		}{}
+	"getFiatRatesForTimestamps": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
+		r := WsFiatRatesForTimestampsReq{}
 		err = json.Unmarshal(req.Params, &r)
 		if err == nil {
-			rv, err = s.getFiatRatesForTimestamps(r.Timestamps, r.Currencies)
+			rv, err = s.getFiatRatesForTimestamps(r.Timestamps, r.Currencies, r.Token)
 		}
 		return
 	},
-	"getFiatRatesTickersList": func(s *WebsocketServer, c *websocketChannel, req *websocketReq) (rv interface{}, err error) {
-		r := struct {
-			Timestamp int64 `json:"timestamp"`
-		}{}
+	"getFiatRatesTickersList": func(s *WebsocketServer, c *websocketChannel, req *WsReq) (rv interface{}, err error) {
+		r := WsFiatRatesTickersListReq{}
 		err = json.Unmarshal(req.Params, &r)
 		if err == nil {
-			rv, err = s.getFiatRatesTickersList(r.Timestamp)
+			rv, err = s.getAvailableVsCurrencies(r.Timestamp, r.Token)
 		}
 		return
 	},
 }
 
-func (s *WebsocketServer) onRequest(c *websocketChannel, req *websocketReq) {
+func (s *WebsocketServer) onRequest(c *websocketChannel, req *WsReq) {
 	var err error
 	var data interface{}
 	defer func() {
@@ -440,7 +430,7 @@ func (s *WebsocketServer) onRequest(c *websocketChannel, req *websocketReq) {
 		}
 		// nil data means no response
 		if data != nil {
-			c.DataOut(&websocketRes{
+			c.DataOut(&WsRes{
 				ID:   req.ID,
 				Data: data,
 			})
@@ -470,20 +460,8 @@ func (s *WebsocketServer) onRequest(c *websocketChannel, req *websocketReq) {
 	}
 }
 
-type accountInfoReq struct {
-	Descriptor     string `json:"descriptor"`
-	Details        string `json:"details"`
-	Tokens         string `json:"tokens"`
-	PageSize       int    `json:"pageSize"`
-	Page           int    `json:"page"`
-	FromHeight     int    `json:"from"`
-	ToHeight       int    `json:"to"`
-	ContractFilter string `json:"contractFilter"`
-	Gap            int    `json:"gap"`
-}
-
-func unmarshalGetAccountInfoRequest(params []byte) (*accountInfoReq, error) {
-	var r accountInfoReq
+func unmarshalGetAccountInfoRequest(params []byte) (*WsAccountInfoReq, error) {
+	var r WsAccountInfoReq
 	err := json.Unmarshal(params, &r)
 	if err != nil {
 		return nil, err
@@ -491,7 +469,7 @@ func unmarshalGetAccountInfoRequest(params []byte) (*accountInfoReq, error) {
 	return &r, nil
 }
 
-func (s *WebsocketServer) getAccountInfo(req *accountInfoReq) (res *api.Address, err error) {
+func (s *WebsocketServer) getAccountInfo(req *WsAccountInfoReq) (res *api.Address, err error) {
 	var opt api.AccountDetails
 	switch req.Details {
 	case "tokens":
@@ -526,14 +504,14 @@ func (s *WebsocketServer) getAccountInfo(req *accountInfoReq) (res *api.Address,
 	if req.PageSize == 0 {
 		req.PageSize = txsOnPage
 	}
-	a, err := s.api.GetXpubAddress(req.Descriptor, req.Page, req.PageSize, opt, &filter, req.Gap)
+	a, err := s.api.GetXpubAddress(req.Descriptor, req.Page, req.PageSize, opt, &filter, req.Gap, strings.ToLower(req.SecondaryCurrency))
 	if err != nil {
-		return s.api.GetAddress(req.Descriptor, req.Page, req.PageSize, opt, &filter)
+		return s.api.GetAddress(req.Descriptor, req.Page, req.PageSize, opt, &filter, strings.ToLower(req.SecondaryCurrency))
 	}
 	return a, nil
 }
 
-func (s *WebsocketServer) getAccountUtxo(descriptor string) (interface{}, error) {
+func (s *WebsocketServer) getAccountUtxo(descriptor string) (api.Utxos, error) {
 	utxo, err := s.api.GetXpubUtxo(descriptor, false, 0)
 	if err != nil {
 		return s.api.GetAddressUtxo(descriptor, false)
@@ -541,7 +519,7 @@ func (s *WebsocketServer) getAccountUtxo(descriptor string) (interface{}, error)
 	return utxo, nil
 }
 
-func (s *WebsocketServer) getTransaction(txid string) (interface{}, error) {
+func (s *WebsocketServer) getTransaction(txid string) (*api.Tx, error) {
 	return s.api.GetTransaction(txid, false, false)
 }
 
@@ -549,30 +527,14 @@ func (s *WebsocketServer) getTransactionSpecific(txid string) (interface{}, erro
 	return s.chain.GetTransactionSpecific(&bchain.Tx{Txid: txid})
 }
 
-func (s *WebsocketServer) getInfo() (interface{}, error) {
+func (s *WebsocketServer) getInfo() (*WsInfoRes, error) {
 	vi := common.GetVersionInfo()
 	bi := s.is.GetBackendInfo()
 	height, hash, err := s.db.GetBestBlock()
 	if err != nil {
 		return nil, err
 	}
-	type backendInfo struct {
-		Version    string      `json:"version,omitempty"`
-		Subversion string      `json:"subversion,omitempty"`
-		Consensus  interface{} `json:"consensus,omitempty"`
-	}
-	type info struct {
-		Name       string      `json:"name"`
-		Shortcut   string      `json:"shortcut"`
-		Decimals   int         `json:"decimals"`
-		Version    string      `json:"version"`
-		BestHeight int         `json:"bestHeight"`
-		BestHash   string      `json:"bestHash"`
-		Block0Hash string      `json:"block0Hash"`
-		Testnet    bool        `json:"testnet"`
-		Backend    backendInfo `json:"backend"`
-	}
-	return &info{
+	return &WsInfoRes{
 		Name:       s.is.Coin,
 		Shortcut:   s.is.CoinShortcut,
 		Decimals:   s.chainParser.AmountDecimals(),
@@ -581,54 +543,55 @@ func (s *WebsocketServer) getInfo() (interface{}, error) {
 		Version:    vi.Version,
 		Block0Hash: s.block0hash,
 		Testnet:    s.chain.IsTestnet(),
-		Backend: backendInfo{
-			Version:    bi.Version,
-			Subversion: bi.Subversion,
-			Consensus:  bi.Consensus,
+		Backend: WsBackendInfo{
+			Version:          bi.Version,
+			Subversion:       bi.Subversion,
+			ConsensusVersion: bi.ConsensusVersion,
+			Consensus:        bi.Consensus,
 		},
 	}, nil
 }
 
-func (s *WebsocketServer) getBlockHash(height int) (interface{}, error) {
+func (s *WebsocketServer) getBlockHash(height int) (*WsBlockHashRes, error) {
 	h, err := s.db.GetBlockHash(uint32(height))
 	if err != nil {
 		return nil, err
 	}
-	type hash struct {
-		Hash string `json:"hash"`
-	}
-	return &hash{
+	return &WsBlockHashRes{
 		Hash: h,
 	}, nil
 }
 
+func (s *WebsocketServer) getBlock(id string, page, pageSize int) (interface{}, error) {
+	block, err := s.api.GetBlock(id, page, pageSize)
+	if err != nil {
+		return nil, err
+	}
+	return block, nil
+}
+
 func (s *WebsocketServer) estimateFee(c *websocketChannel, params []byte) (interface{}, error) {
-	type estimateFeeReq struct {
-		Blocks   []int                  `json:"blocks"`
-		Specific map[string]interface{} `json:"specific"`
-	}
-	type estimateFeeRes struct {
-		FeePerTx   string `json:"feePerTx,omitempty"`
-		FeePerUnit string `json:"feePerUnit,omitempty"`
-		FeeLimit   string `json:"feeLimit,omitempty"`
-	}
-	var r estimateFeeReq
+	var r WsEstimateFeeReq
 	err := json.Unmarshal(params, &r)
 	if err != nil {
 		return nil, err
 	}
-	res := make([]estimateFeeRes, len(r.Blocks))
+	res := make([]WsEstimateFeeRes, len(r.Blocks))
 	if s.chainParser.GetChainType() == bchain.ChainEthereumType {
 		gas, err := s.chain.EthereumTypeEstimateGas(r.Specific)
 		if err != nil {
 			return nil, err
 		}
 		sg := strconv.FormatUint(gas, 10)
-		for i, b := range r.Blocks {
-			fee, err := s.chain.EstimateSmartFee(b, true)
-			if err != nil {
-				return nil, err
-			}
+		b := 1
+		if len(r.Blocks) > 0 {
+			b = r.Blocks[0]
+		}
+		fee, err := s.api.EstimateFee(b, true)
+		if err != nil {
+			return nil, err
+		}
+		for i := range r.Blocks {
 			res[i].FeePerUnit = fee.String()
 			res[i].FeeLimit = sg
 			fee.Mul(&fee, new(big.Int).SetUint64(gas))
@@ -652,7 +615,7 @@ func (s *WebsocketServer) estimateFee(c *websocketChannel, params []byte) (inter
 			}
 		}
 		for i, b := range r.Blocks {
-			fee, err := s.api.BitcoinTypeEstimateFee(b, conservative)
+			fee, err := s.api.EstimateFee(b, conservative)
 			if err != nil {
 				return nil, err
 			}
@@ -677,6 +640,11 @@ func (s *WebsocketServer) sendTransaction(tx string) (res resultSendTransaction,
 	return
 }
 
+func (s *WebsocketServer) getMempoolFilters(r *WsMempoolFiltersReq) (res bchain.MempoolTxidFilterEntries, err error) {
+	res, err = s.mempool.GetTxidFilterEntries(r.ScriptType, r.FromTimestamp)
+	return
+}
+
 type subscriptionResponse struct {
 	Subscribed bool `json:"subscribed"`
 }
@@ -685,7 +653,7 @@ type subscriptionResponseMessage struct {
 	Message    string `json:"message"`
 }
 
-func (s *WebsocketServer) subscribeNewBlock(c *websocketChannel, req *websocketReq) (res interface{}, err error) {
+func (s *WebsocketServer) subscribeNewBlock(c *websocketChannel, req *WsReq) (res interface{}, err error) {
 	s.newBlockSubscriptionsLock.Lock()
 	defer s.newBlockSubscriptionsLock.Unlock()
 	s.newBlockSubscriptions[c] = req.ID
@@ -701,7 +669,7 @@ func (s *WebsocketServer) unsubscribeNewBlock(c *websocketChannel) (res interfac
 	return &subscriptionResponse{false}, nil
 }
 
-func (s *WebsocketServer) subscribeNewTransaction(c *websocketChannel, req *websocketReq) (res interface{}, err error) {
+func (s *WebsocketServer) subscribeNewTransaction(c *websocketChannel, req *WsReq) (res interface{}, err error) {
 	s.newTransactionSubscriptionsLock.Lock()
 	defer s.newTransactionSubscriptionsLock.Unlock()
 	if !s.newTransactionEnabled {
@@ -724,9 +692,7 @@ func (s *WebsocketServer) unsubscribeNewTransaction(c *websocketChannel) (res in
 }
 
 func (s *WebsocketServer) unmarshalAddresses(params []byte) ([]string, error) {
-	r := struct {
-		Addresses []string `json:"addresses"`
-	}{}
+	r := WsSubscribeAddressesReq{}
 	err := json.Unmarshal(params, &r)
 	if err != nil {
 		return nil, err
@@ -760,7 +726,7 @@ func (s *WebsocketServer) doUnsubscribeAddresses(c *websocketChannel) {
 	c.addrDescs = nil
 }
 
-func (s *WebsocketServer) subscribeAddresses(c *websocketChannel, addrDesc []string, req *websocketReq) (res interface{}, err error) {
+func (s *WebsocketServer) subscribeAddresses(c *websocketChannel, addrDesc []string, req *WsReq) (res interface{}, err error) {
 	s.addressSubscriptionsLock.Lock()
 	defer s.addressSubscriptionsLock.Unlock()
 	// unsubscribe all previous subscriptions
@@ -799,16 +765,20 @@ func (s *WebsocketServer) doUnsubscribeFiatRates(c *websocketChannel) {
 			delete(s.fiatRatesSubscriptions, fr)
 		}
 	}
+	delete(s.fiatRatesTokenSubscriptions, c)
 }
 
 // subscribeFiatRates subscribes all FiatRates subscriptions by this channel
-func (s *WebsocketServer) subscribeFiatRates(c *websocketChannel, currency string, req *websocketReq) (res interface{}, err error) {
+func (s *WebsocketServer) subscribeFiatRates(c *websocketChannel, d *WsSubscribeFiatRatesReq, req *WsReq) (res interface{}, err error) {
 	s.fiatRatesSubscriptionsLock.Lock()
 	defer s.fiatRatesSubscriptionsLock.Unlock()
 	// unsubscribe all previous subscriptions
 	s.doUnsubscribeFiatRates(c)
+	currency := d.Currency
 	if currency == "" {
 		currency = allFiatRates
+	} else {
+		currency = strings.ToLower(currency)
 	}
 	as, ok := s.fiatRatesSubscriptions[currency]
 	if !ok {
@@ -816,6 +786,9 @@ func (s *WebsocketServer) subscribeFiatRates(c *websocketChannel, currency strin
 		s.fiatRatesSubscriptions[currency] = as
 	}
 	as[c] = req.ID
+	if len(d.Tokens) != 0 {
+		s.fiatRatesTokenSubscriptions[c] = d.Tokens
+	}
 	s.metrics.WebsocketSubscribes.With((common.Labels{"method": "subscribeFiatRates"})).Set(float64(len(s.fiatRatesSubscriptions)))
 	return &subscriptionResponse{true}, nil
 }
@@ -840,7 +813,7 @@ func (s *WebsocketServer) onNewBlockAsync(hash string, height uint32) {
 		Hash:   hash,
 	}
 	for c, id := range s.newBlockSubscriptions {
-		c.DataOut(&websocketRes{
+		c.DataOut(&WsRes{
 			ID:   id,
 			Data: &data,
 		})
@@ -857,7 +830,7 @@ func (s *WebsocketServer) sendOnNewTx(tx *api.Tx) {
 	s.newTransactionSubscriptionsLock.Lock()
 	defer s.newTransactionSubscriptionsLock.Unlock()
 	for c, id := range s.newTransactionSubscriptions {
-		c.DataOut(&websocketRes{
+		c.DataOut(&WsRes{
 			ID:   id,
 			Data: &tx,
 		})
@@ -885,7 +858,7 @@ func (s *WebsocketServer) sendOnNewTxAddr(stringAddressDescriptor string, tx *ap
 		as, ok := s.addressSubscriptions[stringAddressDescriptor]
 		if ok {
 			for c, id := range as {
-				c.DataOut(&websocketRes{
+				c.DataOut(&WsRes{
 					ID:   id,
 					Data: &data,
 				})
@@ -896,7 +869,7 @@ func (s *WebsocketServer) sendOnNewTxAddr(stringAddressDescriptor string, tx *ap
 }
 
 func (s *WebsocketServer) getNewTxSubscriptions(tx *bchain.MempoolTx) map[string]struct{} {
-	// check if there is any subscription in inputs, outputs and erc20
+	// check if there is any subscription in inputs, outputs and token transfers
 	s.addressSubscriptionsLock.Lock()
 	defer s.addressSubscriptionsLock.Unlock()
 	subscribed := make(map[string]struct{})
@@ -919,8 +892,8 @@ func (s *WebsocketServer) getNewTxSubscriptions(tx *bchain.MempoolTx) map[string
 			}
 		}
 	}
-	for i := range tx.Erc20 {
-		addrDesc, err := s.chainParser.GetAddrDescFromAddress(tx.Erc20[i].From)
+	for i := range tx.TokenTransfers {
+		addrDesc, err := s.chainParser.GetAddrDescFromAddress(tx.TokenTransfers[i].From)
 		if err == nil && len(addrDesc) > 0 {
 			sad := string(addrDesc)
 			as, ok := s.addressSubscriptions[sad]
@@ -928,7 +901,7 @@ func (s *WebsocketServer) getNewTxSubscriptions(tx *bchain.MempoolTx) map[string
 				subscribed[sad] = struct{}{}
 			}
 		}
-		addrDesc, err = s.chainParser.GetAddrDescFromAddress(tx.Erc20[i].To)
+		addrDesc, err = s.chainParser.GetAddrDescFromAddress(tx.TokenTransfers[i].To)
 		if err == nil && len(addrDesc) > 0 {
 			sad := string(addrDesc)
 			as, ok := s.addressSubscriptions[sad]
@@ -960,7 +933,7 @@ func (s *WebsocketServer) OnNewTx(tx *bchain.MempoolTx) {
 	}
 }
 
-func (s *WebsocketServer) broadcastTicker(currency string, rates map[string]float64) {
+func (s *WebsocketServer) broadcastTicker(currency string, rates map[string]float32, ticker *common.CurrencyRatesTicker) {
 	as, ok := s.fiatRatesSubscriptions[currency]
 	if ok && len(as) > 0 {
 		data := struct {
@@ -969,36 +942,60 @@ func (s *WebsocketServer) broadcastTicker(currency string, rates map[string]floa
 			Rates: rates,
 		}
 		for c, id := range as {
-			c.DataOut(&websocketRes{
-				ID:   id,
-				Data: &data,
-			})
+			var tokens []string
+			if ticker != nil {
+				tokens = s.fiatRatesTokenSubscriptions[c]
+			}
+			if len(tokens) > 0 {
+				dataWithTokens := struct {
+					Rates      interface{}        `json:"rates"`
+					TokenRates map[string]float32 `json:"tokenRates,omitempty"`
+				}{
+					Rates:      rates,
+					TokenRates: map[string]float32{},
+				}
+				for _, token := range tokens {
+					rate := ticker.TokenRateInCurrency(token, currency)
+					if rate > 0 {
+						dataWithTokens.TokenRates[token] = rate
+					}
+				}
+				c.DataOut(&WsRes{
+					ID:   id,
+					Data: &dataWithTokens,
+				})
+			} else {
+				c.DataOut(&WsRes{
+					ID:   id,
+					Data: &data,
+				})
+			}
 		}
 		glog.Info("broadcasting new rates for currency ", currency, " to ", len(as), " channels")
 	}
 }
 
 // OnNewFiatRatesTicker is a callback that broadcasts info about fiat rates affecting subscribed currency
-func (s *WebsocketServer) OnNewFiatRatesTicker(ticker *db.CurrencyRatesTicker) {
+func (s *WebsocketServer) OnNewFiatRatesTicker(ticker *common.CurrencyRatesTicker) {
 	s.fiatRatesSubscriptionsLock.Lock()
 	defer s.fiatRatesSubscriptionsLock.Unlock()
 	for currency, rate := range ticker.Rates {
-		s.broadcastTicker(currency, map[string]float64{currency: rate})
+		s.broadcastTicker(currency, map[string]float32{currency: rate}, ticker)
 	}
-	s.broadcastTicker(allFiatRates, ticker.Rates)
+	s.broadcastTicker(allFiatRates, ticker.Rates, nil)
 }
 
-func (s *WebsocketServer) getCurrentFiatRates(currencies []string) (interface{}, error) {
-	ret, err := s.api.GetCurrentFiatRates(currencies)
+func (s *WebsocketServer) getCurrentFiatRates(currencies []string, token string) (*api.FiatTicker, error) {
+	ret, err := s.api.GetCurrentFiatRates(currencies, strings.ToLower(token))
 	return ret, err
 }
 
-func (s *WebsocketServer) getFiatRatesForTimestamps(timestamps []int64, currencies []string) (interface{}, error) {
-	ret, err := s.api.GetFiatRatesForTimestamps(timestamps, currencies)
+func (s *WebsocketServer) getFiatRatesForTimestamps(timestamps []int64, currencies []string, token string) (*api.FiatTickers, error) {
+	ret, err := s.api.GetFiatRatesForTimestamps(timestamps, currencies, strings.ToLower(token))
 	return ret, err
 }
 
-func (s *WebsocketServer) getFiatRatesTickersList(timestamp int64) (interface{}, error) {
-	ret, err := s.api.GetFiatRatesTickersList(timestamp)
+func (s *WebsocketServer) getAvailableVsCurrencies(timestamp int64, token string) (*api.AvailableVsCurrencies, error) {
+	ret, err := s.api.GetAvailableVsCurrencies(timestamp, strings.ToLower(token))
 	return ret, err
 }
